backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
mediapipe==0.10.13
numpy<2.0
scikit-learn>=1.3.0
joblib>=1.3.0
pillow>=10.0.0
onnxruntime>=1.16.0
